requirements.txt
# MeetDocs AI - Python Dependencies

# ============================================
# Browser Automation
# ============================================
selenium==4.16.0
webdriver-manager==4.0.1

# ============================================
# Audio Capture
# ============================================
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.26.2

# ============================================
# Speech-to-Text (Whisper)
# ============================================
openai-whisper==20231117
torch>=2.2.0
torchaudio>=2.2.0

# ============================================
# Translation
# ============================================
googletrans==4.0.0rc1

# ============================================
# AI Summarization (Gemini)
# ============================================
google-generativeai==0.3.2

# ============================================
# Document Export
# ============================================
python-docx==1.1.0

# ============================================
# Configuration Management
# ============================================
python-dotenv==1.0.0

# ============================================
# Testing
# ============================================
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0
hypothesis==6.92.1

# ============================================
# Utilities
# ============================================
# For data classes and type hints (Python 3.10+)
# No additional packages needed - using standard library
